feat: implemented point check on polygon and loading inverted map coordinates

+++ rust/models/map.rs
@@ -3,8 +3,11 @@ use serde::{Deserialize, Serialize};
 /// Represents a 2D map with an area of `width` by `height` tiles.
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Map {
-    pub width: f32,
-    pub height: f32,
+    // the width of the map
+    pub width: f64,
+    // the height of the map
+    pub height: f64,
+    // the layers that exist within the map
     pub layers: Vec<Layer>
 }
 
@@ -25,20 +28,22 @@ pub struct Object {
     // used to mark an object as an ellipse
     pub ellipse: Option<bool>,
     // represents the height of the object
-    pub height: f32,
+    pub height: f64,
     // represents the width of the object
-    pub width: f32,
+    pub width: f64,
     // if the object is a polygon, this represents the points that make up the polygon
     pub polygon: Option<Vec<Point>>,
     // represents the x coordinate of the object within the map
-    pub x: f32,
+    pub x: f64,
     // represents the y coordinate of the object within the map
-    pub y: f32
+    pub y: f64
 }
 
-/// Represents a point on a polygon or a polyline in a Tiled map.
-#[derive(Serialize, Deserialize, Debug)]
+/// Represents a point on a polygon or a polyline with x and y coordinates relative to the object
+#[derive(Clone, Copy, Serialize, Deserialize, Debug)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32
+    // the x coordinate of the point
+    pub x: f64,
+    // the y coordinate of the point
+    pub y: f64
 }
